requirements.txt
# Core Data Science & ML
numpy>=1.21.0,<1.25.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Audio Processing
librosa>=0.9.0
soundfile>=0.10.0
audioread>=2.1.0

# Machine Learning - TensorFlow with protobuf fix
tensorflow>=2.10.0,<2.16.0
protobuf>=3.19.0,<3.20.0
tensorflow-io>=0.24.0
tensorboard>=2.10.0

# Additional ML utilities
joblib>=1.1.0

# Web App & UI
streamlit>=1.28.0
streamlit-webrtc>=0.45.0

# QR Code Generation
qrcode[pil]>=7.3.0
Pillow>=9.0.0

# Jupyter & Development
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Utilities
tqdm>=4.64.0
python-dotenv>=0.19.0

# Optional: for production deployment
# gunicorn>=20.1.0
# uvicorn>=0.18.0